pyproject.toml
[project]
name = "future-django"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "django>=6.0.3",
    "django-ninja>=1.6.2",
    "pydantic>=2.12.5",
    "pytest-django>=4.12.0",
    "python-ulid>=3.1.0",
    "ruff>=0.15.8",
    "svcs>=25.1.0",
]

[dependency-groups]
dev = [
    "django-stubs>=6.0.2",
    "django-types>=0.23.0",
    "pyrefly>=0.59.0",
    "pyright>=1.1.408",
]

[tool.pyright]
pythonVersion = "3.14"
include = ["src"]
venvPath = "."
venv = ".venv"
